requirements.txt
dj-database-url==0.4.2
Django==1.11.7
djangorestframework==3.7.3
gunicorn==19.7.1
jsonfield==2.0.2
numpy==1.13.3
olefile==0.44
opencv-python==3.3.0.10
Pillow==4.3.0
psycopg2==2.7.3.2
pymongo==3.5.1
pytz==2017.3
ruamel.appconfig==0.5.3
ruamel.std.argparse==0.8.1
whitenoise==3.3.1
yaml-1.3==0.1.0
